Change of plans: reinitialize budgeting app as terminal application
I decided to change the direction of the application. Priviously, I
thought to create a web/mobile application which comminucate via a
remote server. Now, I plan to create a simpler terminal application
which sources a local data file.

Most of the code in `index.coffee` is just a result of playing around
with the blessed javascript terminal library.